MainMenu done
I completed creation of Main menu state.
It will now show controls for the game, about the game, it can exit the
game and start the game. On On F4 you can exit whole game from main menu
as well as from main game.
On ESC you can go back from controls/about.

Also there's an UPLOAD PHOTO button and if you click on it then you will
choose a photo that will be edited in the main game.